+++ packages/ricos-content/web/src/converters/html/fromHtml/ck-editor/postprocess.spec.ts
@@ -1,8 +1,8 @@
 import { RichContent } from 'ricos-schema';
-import { compare } from '../../../comparision/compare';
-import preprocess from './preprocess';
+import { compare } from '../../../../comparision/compare';
+import postprocess from './postprocess';
 
-describe('toDraft preprocess', () => {
+describe('CKEditor postprocess', () => {
   it('<li><p><p></li> => <li><p></li>', () => {
     const content: RichContent = RichContent.fromJSON({
       nodes: [
@@ -106,7 +106,7 @@ describe('toDraft preprocess', () => {
         },
       ],
     });
-    const actual = preprocess(content);
+    const actual = postprocess(content);
 
     expect(RichContent.fromJSON(actual)).toStrictEqual(expected);
   });
@@ -372,7 +372,7 @@ describe('toDraft preprocess', () => {
         },
       ],
     });
-    const actual = preprocess(content);
+    const actual = postprocess(content);
     expect(compare(RichContent.fromJSON(actual), expected, { ignoredKeys: ['id'] })).toEqual({});
   });
 });